feat(20-01): bump file format version to 3 and update constants
Bumped FILE_FORMAT_VERSION and V2_FORMAT_VERSION from 2 to 3 to indicate
the new schema_version format (u32 instead of u64). Updated SCHEMA_VERSION
header size from 8 to 4 bytes and changed DEFAULT_SCHEMA_VERSION from u64 to u32.

Changes:
- constants.rs: FILE_FORMAT_VERSION = 3, SCHEMA_VERSION size = 4, DEFAULT_SCHEMA_VERSION = u32
- v2/mod.rs: V2_FORMAT_VERSION = 3 with comment about v3 schema format
- persistent_header.rs: Removed cast to u32 (no longer needed)
- types/file_header.rs: Updated schema_version to u32, added reserved field, updated checksum calculation

+++ sqlitegraph/src/backend/native/persistent_header.rs
@@ -80,7 +80,7 @@ impl PersistentHeaderV2 {
             flags: crate::backend::native::constants::DEFAULT_FEATURE_FLAGS,
             node_count: 0,
             edge_count: 0,
-            schema_version: crate::backend::native::constants::DEFAULT_SCHEMA_VERSION as u32,
+            schema_version: crate::backend::native::constants::DEFAULT_SCHEMA_VERSION,
             reserved: 0,
             node_data_offset: crate::backend::native::constants::HEADER_SIZE,
             edge_data_offset: crate::backend::native::constants::HEADER_SIZE,
